DEF CON 24 Practicum/Demo Eclipse workspace
This is the eclipse workspace. It contains the code for the Practicum/Demo. 
Download and extract this file. 
When you start eclipse, point to the workspace directory that was in the zip. 
You will also need to download and setup Tomcat 7 and JDK 1.8
See the workshop slides for more information on how to set everything up. It has links to tutorials.